drop import subcommand
